Build 280 Tristate change to checkboxes complete Reorganization of checkboxes Hash calculation cleaned up Adjustments to lastfile.txt to utilize CheckState for tristate boxes Change the ! to # in flags so the filename is created correctly

+++ DW3Randomizer/classes/initialize.cs
@@ -10,7 +10,7 @@ namespace DW3Randomizer
     {
         public void setrandomization(ref Random r1, ref Random randomFlagIncrement, ref Random randomCosmeticIncrement, string txtSeed, int buildnumber, bool genCompareFile)
         {
-            r1 = new Random((int)long.Parse(txtSeed));
+            r1 = new Random(int.Parse(txtSeed));
             randomFlagIncrement = new Random(int.Parse(txtSeed));
             randomCosmeticIncrement = new Random(int.Parse(txtSeed));
